Refactor the logging package to trace
