+++ tomcat-gcp-lib/src/test/java/com/google/cloud/runtimes/tomcat/trace/TraceValveTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.runtimes.tomcat.logging;
+package com.google.cloud.runtimes.tomcat.trace;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
